+++ backend/src/tests/Functional/Core/Api/App/CharControllerTest.php
@@ -19,6 +19,10 @@ class CharControllerTest extends WebTestCase
      */
     private $repoFactory;
 
+    private $app0Id;
+
+    private $appId;
+
     public function setUp()
     {
         $this->helper = new Helper();
@@ -27,28 +31,42 @@ class CharControllerTest extends WebTestCase
 
     public function testMainV1403()
     {
-        $response = $this->runApp('GET', '/api/app/v1/main/123');
-        $this->assertEquals(403, $response->getStatusCode());
+        $response1 = $this->runApp('GET', '/api/app/v1/main/123');
+        $this->assertEquals(403, $response1->getStatusCode());
+
+        $this->setUpDb();
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->app0Id.':s1')];
+        $response2 = $this->runApp('GET', '/api/app/v1/main/123', null, $headers);
+        $this->assertEquals(403, $response2->getStatusCode());
     }
 
     public function testMainV1404()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $this->setUpDb();
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/main/123', null, $headers);
 
         $this->assertEquals(404, $response->getStatusCode());
         $this->assertEquals('Not Found', $response->getReasonPhrase());
     }
 
+    public function testMainV2403()
+    {
+        $response1 = $this->runApp('GET', '/api/app/v2/main/123');
+        $this->assertEquals(403, $response1->getStatusCode());
+
+        $this->setUpDb();
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->app0Id.':s1')];
+        $response2 = $this->runApp('GET', '/api/app/v2/main/123', null, $headers);
+        $this->assertEquals(403, $response2->getStatusCode());
+    }
+
     public function testMainV2404()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $this->setUpDb();
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response = $this->runApp('GET', '/api/app/v2/main/123', null, $headers);
 
         $this->assertEquals(404, $response->getStatusCode());
@@ -57,14 +75,12 @@ class CharControllerTest extends WebTestCase
 
     public function testMainV1204()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
-
+        $this->setUpDb();
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $char->setMain(false);
         $this->helper->getEm()->flush();
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/main/123', null, $headers);
 
         $this->assertEquals(204, $response->getStatusCode());
@@ -72,12 +88,11 @@ class CharControllerTest extends WebTestCase
 
     public function testMainV1200()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $this->setUpDb();
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $this->helper->addCharacterToPlayer('C2', 456, $char->getPlayer());
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response1 = $this->runApp('GET', '/api/app/v1/main/123', null, $headers);
         $response2 = $this->runApp('GET', '/api/app/v1/main/456', null, $headers);
 
@@ -105,14 +120,18 @@ class CharControllerTest extends WebTestCase
     {
         $response = $this->runApp('GET', '/api/app/v1/characters/123');
         $this->assertEquals(403, $response->getStatusCode());
+
+        $this->setUpDb();
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->app0Id.':s1')];
+        $response = $this->runApp('GET', '/api/app/v1/characters/123', null, $headers);
+        $this->assertEquals(403, $response->getStatusCode());
     }
 
     public function testCharactersV1404()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $this->setUpDb();
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response = $this->runApp('GET', '/api/app/v1/characters/123', null, $headers);
 
         $this->assertEquals(404, $response->getStatusCode());
@@ -121,12 +140,11 @@ class CharControllerTest extends WebTestCase
 
     public function testCharactersV1200()
     {
-        $this->helper->emptyDb();
-        $aid = $this->helper->addApp('A1', 's1', [Role::APP])->getId();
+        $this->setUpDb();
         $char = $this->helper->addCharacterMain('C1', 123, [Role::USER]);
         $this->helper->addCharacterToPlayer('C2', 456, $char->getPlayer());
 
-        $headers = ['Authorization' => 'Bearer '.base64_encode($aid.':s1')];
+        $headers = ['Authorization' => 'Bearer '.base64_encode($this->appId.':s1')];
         $response1 = $this->runApp('GET', '/api/app/v1/characters/123', null, $headers);
         $response2 = $this->runApp('GET', '/api/app/v1/characters/456', null, $headers);
 
@@ -154,5 +172,12 @@ class CharControllerTest extends WebTestCase
             ]],
             $body1
         );
+    }
+
+    private function setUpDb()
+    {
+        $this->helper->emptyDb();
+        $this->app0Id = $this->helper->addApp('A0', 's0', [Role::APP])->getId();
+        $this->appId = $this->helper->addApp('A1', 's1', [Role::APP, Role::APP_CHARS])->getId();
     }
 }