+++ CHANGELOG.md
@@ -7,6 +7,7 @@ dd mm yyyy
 - Added a script that creates a distribution.
 - Added ESI proxy endpoint for apps.
 - Added app endpoint that combines the player groups, corp groups and alliance groups endpoints.
+- Implemented more fine grained permissions for apps (new roles app-groups and app-chars).
 
 - TODO ESI cache